seperate .py file for preprocessing
